pyproject.toml
[project]
name = "fastapi-server-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiomysql>=0.2.0",
    "fastapi[standard]>=0.118.0",
    "greenlet>=3.2.4",
    "mypy>=1.18.2",
    "orjson>=3.11.3",
    "passlib[bcrypt]>=1.7.4",
    "pre-commit>=4.3.0",
    "pydantic-settings>=2.11.0",
    "pytest>=8.4.2",
    "python-jose[cryptography]>=3.5.0",
    "ruff>=0.13.2",
    "sqlalchemy>=2.0.43",
]

[tool.ruff]
line-length = 80

[tool.ruff.lint]
select = [
    "A",
    "ASYNC",
    "C90",
    "COM",
    "ERA001",
    "F",
    "FAST",
    "I",
    "N",
    "S",
    "T20",
]

ignore = [
    "COM812",
]

[[tool.mypy.overrides]]
module = [
    "jose.*",
    "passlib.context",
]
ignore_missing_imports = true
